renamed data_sources package and moved StubDataSource there
