+++ app/src/main/kotlin/com/igor_shaula/api_polling/data_layer/data_sources/NetworkDataSource.kt
@@ -1,8 +1,8 @@
-package com.igor_shaula.api_polling.data_layer.network_data_source
+package com.igor_shaula.api_polling.data_layer.data_sources
 
 import com.igor_shaula.api_polling.data_layer.VehicleDetailsRecord
 import com.igor_shaula.api_polling.data_layer.VehicleRecord
-import com.igor_shaula.api_polling.data_layer.network_data_source.retrofit.VehicleRetrofitNetworkServiceImpl
+import com.igor_shaula.api_polling.data_layer.data_sources.retrofit.VehicleRetrofitNetworkServiceImpl
 
 class NetworkDataSource {
