+++ app/src/main/kotlin/com/igor_shaula/api_polling/data_layer/data_sources/api_stub/VehicleStubNetworkServiceImpl.kt
@@ -1,10 +1,10 @@
-package com.igor_shaula.api_polling.data_layer.network_data_source.api_stub
+package com.igor_shaula.api_polling.data_layer.data_sources.api_stub
 
 import com.igor_shaula.api_polling.data_layer.TARGET_LATITUDE
 import com.igor_shaula.api_polling.data_layer.TARGET_LONGITUDE
-import com.igor_shaula.api_polling.data_layer.network_data_source.LocationModel
-import com.igor_shaula.api_polling.data_layer.network_data_source.VehicleDetailsModel
-import com.igor_shaula.api_polling.data_layer.network_data_source.VehicleModel
+import com.igor_shaula.api_polling.data_layer.data_sources.LocationModel
+import com.igor_shaula.api_polling.data_layer.data_sources.VehicleDetailsModel
+import com.igor_shaula.api_polling.data_layer.data_sources.VehicleModel
 import com.igor_shaula.stub_data_generator.StubVehicleGenerator
 import kotlin.random.Random
 import kotlin.random.nextInt
